Move NTLM code into `auth` package
