+++ auth/ntlm_test.go
@@ -1,4 +1,4 @@
-package lfs
+package auth
 
 import (
 	"bytes"
